Rewrite some on top of the session history rewrite
This is not complete, but mainly attempts to tackle the "Updating the Navigation object" section, in particular to get rid of the TODO around the navigationType set when firing currententrychange. A notable result is clarifying the relative timing of currententrychange/dispose vs. pageshow vs. hashchange/popstate.

Other sections remain on top of the current HTML spec, or on top of older revisions of the session history rewrite.

This also replaces the abbrevation "AHE" (AppHistoryEntry) with "NHE" (NavigationHistoryEntry) inside the "update the entries" algorithm.