Rebase more on top of session history rewrite
Similar to the previous commit, this attempts to tackle a specific area, not the whole rebase project (which is tracked in #221). It focuses on traversal, in particular integration with the new versions of "apply the history step" and "traverse the history by a delta".

This removes a TODO in the former definition of "perform a navigation API traversal".