Revamp intercepted-navigation scroll handling
This changes the default behavior of intercepted navigations with respect to scrolling in the following ways:

* Scroll restoration is performed on reloads, not just traverses.
* For pushes and replaces, we either scroll to the top or scroll to the fragment.

This change requires some updated API surface: the scrollRestoration option to intercept() has become scroll, and the navigateEvent.restoreScroll() method has become navigateEvent.scroll(). The latter method now has more capabilities, working to give browser-default-like behavior for "push", "replace", and "reload" in addition to "traverse".

Similar to before, all of this can be opted out of by using scroll: "manual".

Closes #237 by giving an easy default behavior for SPA navigations with hashes.

Closes #231 by making "push", "replace", and "reload" behave by default in an MPA-like manner with respect to scroll position, just like "traverse" does.