Allow traversals to be cancelable
This follows the explainer update in a430943e9e82c9288120f2201dabf588f126a273. In the process, this closes #241 by rewriting the appropriate parts of the spec.

This mostly solves #32, but we will leave that open until we add the explainer and spec for consumable sticky activation, which will complete the feature.