Add an explainer for navigateEvent.redirect()
Part of #124. This replaces the previous idea of navigation.transition.rollback(), with something that is only scoped to the { commit: "after-transition" } case for now. A future extension to { commit: "immediate" } is gestured at.

Also remove the explainer section for rollback, since we've had fewer requests for it. We can keep #124 open to discuss it though.